Implement A-Packages - Create class file in package for each class used - Split subclasses into their own class file
