+++ src/main/java/duke/DukeConstants.java
@@ -1,0 +1,9 @@
+package duke;
+
+public final class DukeConstants {
+    private DukeConstants() {
+
+    }
+    public static final String KEY_SEPARATOR = "//";
+    public static final String FILENAME = "bin/tasks.txt";
+}
