+++ src/main/java/duke/DukeException.java
@@ -1,0 +1,18 @@
+package duke;
+
+public class DukeException extends Exception {
+
+    public static final String UNRECOGNISED_COMMAND = "i don't know what that means!";
+    public static final String OUT_OF_RANGE = "there's nothing there :<";
+    public static final String MISSING_INDEX = "pls enter an index!";
+    public static final String MISSING_DESCRIPTION = "description cannot be empty!!";
+    public static final String MISSING_DATE= "date cannot be empty!";
+    public static final String WRONG_FORMAT = "wrong format!";
+    public static final String WRONG_FORMAT_DATE = "wrong date format! pls re-enter using yyyy-mm-dd";
+
+    public DukeException(String message) {
+        super(message);
+    }
+
+
+}
