+++ src/main/java/duke/ListCommand.java
@@ -1,0 +1,18 @@
+package duke;
+
+public class ListCommand extends Command {
+    public ListCommand() {
+
+    }
+
+    @Override
+    public TaskList execute(TaskList taskList, Ui ui, Storage storage) {
+        ui.printTasks(taskList);
+        return taskList;
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
