+++ src/main/java/duke/MarkCommand.java
@@ -1,0 +1,23 @@
+package duke;
+
+public class MarkCommand extends Command {
+    int index;
+
+    public MarkCommand(int index) {
+        this.index = index;
+    }
+
+    @Override
+    public TaskList execute(TaskList taskList, Ui ui, Storage storage) {
+        taskList.markTask(index);
+        Task temp = taskList.getTask(index);
+        ui.displayTask(ui.MARKED, temp);
+        storage.save(taskList, ui);
+        return taskList;
+    }
+
+    @Override
+    public boolean isExit() {
+        return false;
+    }
+}
